PR comments

+++ src/interfaces/ILBPStrategyBasic.sol
@@ -14,7 +14,7 @@ interface ILBPStrategyBasic is IDistributionContract {
     error MigrationNotAllowed(uint256 migrationBlock, uint256 currentBlock);
 
     /// @notice Error thrown when the token split is too high
-    error TokenSplitTooHigh(uint16 tokenSplit);
+    error TokenSplitTooHigh(uint24 tokenSplit);
 
     /// @notice Error thrown when the tick spacing is greater than the max tick spacing or less than the min tick spacing
     error InvalidTickSpacing(int24 tickSpacing);
